requirements.txt
colorama          0.4.6
contourpy         1.3.2
cycler            0.12.1
filelock          3.20.0
fonttools         4.61.1
fsspec            2025.12.0
Jinja2            3.1.3
joblib            1.5.3
kiwisolver        1.4.9
MarkupSafe        3.0.2
matplotlib        3.10.8
mpmath            1.3.0
networkx          3.4.2
numpy             2.2.6
packaging         26.0
pillow            12.0.0
pip               23.0.1
pyparsing         3.3.2
python-dateutil   2.9.0.post0
scikit-learn      1.7.2
scipy             1.15.3
setuptools        65.5.0
six               1.17.0
sympy             1.13.1
threadpoolctl     3.6.0
torch             2.5.1+cu121
torchaudio        2.5.1+cu121
torchvision       0.20.1+cu121
tqdm              4.67.3
typing_extensions 4.9.0